dragging pins works (almost)
there is an ugly bug when start dragging the pin the view shifts, it
shifts back as soon as you drag it. This is because
imageView.setPanEnabled moves the view to the center of the screen
(because of dave morrisey). Annoying as f**k but we'll find a way around
it I'm sure